+++ ActiveDataProvider.php
@@ -158,4 +158,13 @@ class ActiveDataProvider extends \yii\data\ActiveDataProvider
             return array_keys($models);
         }
     }
+
+    /**
+     * @inheritdoc
+     */
+    public function refresh()
+    {
+        parent::refresh();
+        $this->_queryResults = null;
+    }
 }